feat(apps): rip out Apps v1 — editor, routes, agent tools, and UI
RealAdvisor is 58/58 on v2 with v1 writes sealed (#815); no other workspace
clears the activity bar (one hobby workspace, apps untouched since Aug 16).
Data is untouched: MakoApp documents, entity_versions, and the migration CLI
all stay, so any workspace can still be migrated later.

Removed
- API: v1 CRUD routes (routes/apps.ts), server-app-tools (all 21 tools),
  persist-app-draft; registrations in unified agent, MCP server, and
  bridge inventory; APP_V1_ONLY set + isolateAppToolFamily (v1 tabs no
  longer exist); v1 names out of APP_MODE list; App Mode prompt is v2-only.
- Shared package: the 20 dead v1 capability definitions (open_app/run_app/
  app_set_preview* client caps remain — run_app still powers headless MCP
  verify via preview-tools).
- App: AppsExplorer/AppFileEditor/AppBindingEditor/AppRenderer, appStore,
  app-runtime shell/agent-tools/binding-preview, acp-app-focus; v1 tab
  kinds, deep links (/a/:id), breadcrumbs, command-palette search, realtime
  refetch, client tool dispatch, desktop ACP handlers; the v1 Apps rail
  entry (an old persisted "apps" pane now lands on Apps v2).

Kept (the 45 live v1 public-share links keep working)
- public-share routes + app-preview routes, PublicAppViewer/AppPreviewPage,
  app-runtime preview/duckdb/app-location (AppEntity type moved to
  app-runtime/app-entity.ts), app-version + binding-materialization
  services and their Inngest refresh (public links keep fresh data).

Tests: v1 assertions swapped to surviving tools; UrlSync uiStore mock got
getState (pre-existing gap); persist-app-draft test dropped from the chain.
