+++ api/package.json
@@ -14,7 +14,7 @@
     "migrate:workspaces": "tsx src/migrations/migrate-to-workspaces.ts",
     "openapi:generate": "tsx src/openapi/generate-cli.ts",
     "tools:measure": "tsx src/scripts/measure-tool-tokens.ts",
-    "test": "tsx src/openapi/core.test.ts && tsx src/openapi/openapi.test.ts && tsx src/dbt/warm-dir-cache.test.ts && tsx src/connectors/stripe/connector.test.ts && tsx src/connectors/close/connector.test.ts && tsx src/connectors/pandadoc/connector.test.ts && tsx src/connectors/claap/connector.test.ts && tsx src/connectors/calendly/connector.test.ts && tsx src/connectors/wise/connector.test.ts && tsx src/connectors/incremental-capability.test.ts && tsx src/sync-cdc/idempotency.test.ts && tsx src/services/flow-triggers.service.test.ts && tsx src/agent-lib/context/compaction.test.ts && tsx src/agent-lib/tool-input-repair.test.ts && tsx src/agent-lib/stream-error.test.ts && tsx src/services/safe-fetch.service.test.ts && tsx src/services/mcp-client.service.test.ts && tsx src/services/agent-thread.service.test.ts && tsx src/services/agent-stream.service.test.ts && tsx src/mcp/mako-mcp-server.test.ts && tsx src/services/app-preview-token.service.test.ts && tsx src/utils/message-sanitizer.test.ts && tsx src/utils/mongoose-dirty-paths.test.ts && tsx src/services/persist-app-draft.test.ts && tsx src/agent-lib/tools/check-query-status-poll.test.ts && tsx src/agent-lib/tools/str-replace.test.ts && tsx src/agent-lib/tools/self-directive-tool.test.ts && tsx src/agent-lib/tools/universal-tools.test.ts && tsx src/agents/modes/derive-mode-state.test.ts && tsx src/agents/modes/tool-working-set.test.ts && tsx src/agent-lib/tools/tool-discovery-tools.test.ts && tsx src/auth/dev-login.test.ts && tsx src/auth/signup-attribution.test.ts && tsx src/services/parquet-build.readonly.test.ts && tsx src/services/kernel-token.service.test.ts && tsx src/apps-v2/sandbox/never-commit.test.ts && tsx src/notebooks/store/filesystem-store.test.ts && tsx src/services/kernel-provider/kernel-gateway-client.test.ts && tsx src/services/kernel-session-store.test.ts && vitest run --config vitest.apps-v2.config.ts",
+    "test": "tsx src/openapi/core.test.ts && tsx src/openapi/openapi.test.ts && tsx src/dbt/warm-dir-cache.test.ts && tsx src/connectors/stripe/connector.test.ts && tsx src/connectors/close/connector.test.ts && tsx src/connectors/pandadoc/connector.test.ts && tsx src/connectors/claap/connector.test.ts && tsx src/connectors/calendly/connector.test.ts && tsx src/connectors/wise/connector.test.ts && tsx src/connectors/incremental-capability.test.ts && tsx src/sync-cdc/idempotency.test.ts && tsx src/services/flow-triggers.service.test.ts && tsx src/agent-lib/context/compaction.test.ts && tsx src/agent-lib/tool-input-repair.test.ts && tsx src/agent-lib/stream-error.test.ts && tsx src/services/safe-fetch.service.test.ts && tsx src/services/mcp-client.service.test.ts && tsx src/services/agent-thread.service.test.ts && tsx src/services/agent-stream.service.test.ts && tsx src/mcp/mako-mcp-server.test.ts && tsx src/services/app-preview-token.service.test.ts && tsx src/utils/message-sanitizer.test.ts && tsx src/utils/mongoose-dirty-paths.test.ts && tsx src/agent-lib/tools/check-query-status-poll.test.ts && tsx src/agent-lib/tools/str-replace.test.ts && tsx src/agent-lib/tools/self-directive-tool.test.ts && tsx src/agent-lib/tools/universal-tools.test.ts && tsx src/agents/modes/derive-mode-state.test.ts && tsx src/agents/modes/tool-working-set.test.ts && tsx src/agent-lib/tools/tool-discovery-tools.test.ts && tsx src/auth/dev-login.test.ts && tsx src/auth/signup-attribution.test.ts && tsx src/services/parquet-build.readonly.test.ts && tsx src/services/kernel-token.service.test.ts && tsx src/apps-v2/sandbox/never-commit.test.ts && tsx src/notebooks/store/filesystem-store.test.ts && tsx src/services/kernel-provider/kernel-gateway-client.test.ts && tsx src/services/kernel-session-store.test.ts && vitest run --config vitest.apps-v2.config.ts",
     "test:dbt": "vitest run",
     "test:dbt:watch": "vitest",
     "test:destinations": "vitest run --config vitest.destinations.config.ts",
