+++ api/src/routes/register-routes.ts
@@ -26,7 +26,6 @@ import { usageRoutes } from "./usage";
 import { billingRoutes } from "./billing";
 import { stripeWebhookRoutes } from "./stripe-webhook";
 import { dashboardRoutes } from "./dashboards";
-import { appRoutes } from "./apps";
 import { notebookDataRoutes } from "./notebook-data";
 import { notebookRoutes } from "./notebooks";
 import { notebookSessionRoutes } from "./notebook-sessions";
@@ -98,7 +97,6 @@ export function registerApiRoutes(app: OpenAPIHono<AuthEnv>): void {
   app.route("/api/workspaces/:workspaceId/usage", usageRoutes);
   app.route("/api/workspaces/:workspaceId/billing", billingRoutes);
   app.route("/api/workspaces/:workspaceId/dashboards", dashboardRoutes);
-  app.route("/api/workspaces/:workspaceId/apps", appRoutes);
   app.route("/api/workspaces/:workspaceId/notebook", notebookDataRoutes);
   app.route("/api/workspaces/:workspaceId/notebooks", notebookRoutes);
   app.route("/api/workspaces/:workspaceId/notebooks", notebookSessionRoutes);
