--- app/src/components/AppBindingEditor.tsx
@@ -1,468 +1,0 @@
-import { useCallback, useEffect, useState } from "react";
-import {
-  Box,
-  MenuItem,
-  Select,
-  Typography,
-  ToggleButtonGroup,
-  ToggleButton,
-  Tooltip,
-} from "@mui/material";
-import { Info as InfoIcon } from "lucide-react";
-import { Panel, PanelGroup, PanelResizeHandle } from "react-resizable-panels";
-import { containsDbtSchemaToken } from "@mako/schemas";
-import { useWorkspace } from "../contexts/workspace-context";
-import { useAppStore } from "../store/appStore";
-import { useConsoleStore } from "../store/consoleStore";
-import { useDbtStore } from "../store/dbtStore";
-import { previewParquetArtifact } from "../lib/parquet-preview";
-import Console from "./Console";
-import ResultsTable from "./ResultsTable";
-import EntityBreadcrumbs from "./EntityBreadcrumbs";
-import EntityLoadErrorState, {
-  EntityLoadingState,
-} from "./EntityLoadErrorState";
-import { missingEntityError } from "../lib/entity-labels";
-import DataSourceMaterializationControls, {
-  type MaterializationHistoryItem,
-} from "./DataSourceMaterializationControls";
-import {
-  defaultMaterializationSchedule,
-  type MaterializationScheduleValue,
-} from "../lib/materializationSchedule";
-
-interface PreviewResult {
-  results: Record<string, unknown>[];
-  executedAt: string;
-  resultCount: number;
-  executionTime?: number;
-  fields?: Array<{ name?: string; originalName?: string } | string>;
-}
-
-function bindingFilePath(name: string, language: string): string {
-  const ext = language === "sql" ? "sql" : "js";
-  return `${name.replace(/[^a-zA-Z0-9_-]/g, "_")}.${ext}`;
-}
-
-/**
- * Data source inspector. Reuses the polished `Console` component (run / save /
- * connection / editor) so app + dashboard data sources look and behave exactly
- * like a saved console, with a small set of data-source-specific controls
- * (materialization) injected into the toolbar.
- */
-export default function AppBindingEditor({
-  tabId,
-  appId,
-  bindingId,
-}: {
-  tabId: string;
-  appId: string;
-  bindingId: string;
-}) {
-  const { currentWorkspace } = useWorkspace();
-  const workspaceId = currentWorkspace?.id;
-
-  const appEntity = useAppStore(s => s.openApps[appId]);
-  const appLoadError = useAppStore(s => s.openAppErrors[appId]);
-  const fetchApp = useAppStore(s => s.fetchApp);
-  const updateBinding = useAppStore(s => s.updateBinding);
-  const persistApp = useAppStore(s => s.persistApp);
-  const materializeBinding = useAppStore(s => s.materializeBinding);
-  const executeQuery = useConsoleStore(s => s.executeQuery);
-  const updateTabTitle = useConsoleStore(s => s.updateTitle);
-
-  const [preview, setPreview] = useState<PreviewResult | null>(null);
-  const [viewMode, setViewMode] = useState<"table" | "json" | "chart">("table");
-  const [running, setRunning] = useState(false);
-  const [materializing, setMaterializing] = useState(false);
-  const [previewingSnapshot, setPreviewingSnapshot] = useState(false);
-
-  useEffect(() => {
-    if (!appEntity && workspaceId) void fetchApp(workspaceId, appId);
-  }, [appEntity, workspaceId, appId, fetchApp]);
-
-  const binding = appEntity?.dataBindings.find(b => b.id === bindingId);
-
-  // Keep the tab title in sync with the binding name (e.g. when the tab was
-  // opened from a deep link with a placeholder title).
-  const bindingName = binding?.name;
-  useEffect(() => {
-    if (bindingName) updateTabTitle(tabId, bindingName);
-  }, [bindingName, tabId, updateTabTitle]);
-
-  const resolveDbtCodeForPreview = useAppStore(s => s.resolveDbtCodeForPreview);
-  const bindingDbtProjectId = binding?.dbtProjectId;
-
-  // dbt link: lets the user connect this binding to a workspace dbt project
-  // from the UI (previously agent-tool only). Linking enables the
-  // {{ dbt_schema }} token in the query and, with it, the app preview's
-  // environment (schema) switcher — including for materialized (parquet)
-  // bindings, which serve a live row-capped run while an override is active.
-  const dbtProjects = useDbtStore(s => s.projects);
-  const dbtProjectsLoaded = useDbtStore(s => s.projectsLoaded);
-  const fetchDbtProjects = useDbtStore(s => s.fetchProjects);
-  useEffect(() => {
-    if (workspaceId && !dbtProjectsLoaded) void fetchDbtProjects(workspaceId);
-  }, [workspaceId, dbtProjectsLoaded, fetchDbtProjects]);
-
-  const handleDbtLinkChange = useCallback(
-    (projectId: string) => {
-      if (!workspaceId) return;
-      updateBinding(appId, bindingId, {
-        dbtProjectId: projectId || undefined,
-      });
-      void persistApp(workspaceId, appId);
-    },
-    [workspaceId, appId, bindingId, updateBinding, persistApp],
-  );
-
-  const handleExecute = useCallback(
-    async (content: string, connectionId?: string, databaseId?: string) => {
-      if (!workspaceId || !connectionId) return;
-      setRunning(true);
-      try {
-        // dbt-linked bindings: resolve {{ dbt_schema }} against the app's
-        // preview environment (override or prod default) before running.
-        const resolved = await resolveDbtCodeForPreview(
-          workspaceId,
-          appId,
-          bindingDbtProjectId,
-          content,
-        );
-        const res = await executeQuery(workspaceId, connectionId, resolved, {
-          databaseId,
-          databaseName: binding?.databaseName,
-        });
-        if (res.success) {
-          setPreview({
-            results: res.rows || [],
-            executedAt: new Date().toISOString(),
-            resultCount: res.rows?.length ?? 0,
-            executionTime: res.executionTime,
-            fields: res.fields,
-          });
-        } else {
-          setPreview({
-            results: [{ error: res.error || "Query failed" }],
-            executedAt: new Date().toISOString(),
-            resultCount: 0,
-          });
-        }
-      } finally {
-        setRunning(false);
-      }
-    },
-    [
-      workspaceId,
-      appId,
-      executeQuery,
-      binding?.databaseName,
-      bindingDbtProjectId,
-      resolveDbtCodeForPreview,
-    ],
-  );
-
-  const handleSave = useCallback(
-    async (content: string) => {
-      if (!workspaceId) return false;
-      updateBinding(appId, bindingId, { code: content });
-      await persistApp(workspaceId, appId);
-      return true;
-    },
-    [workspaceId, appId, bindingId, updateBinding, persistApp],
-  );
-
-  const handleMaterialize = useCallback(async () => {
-    if (!workspaceId) return;
-    setMaterializing(true);
-    // Explicit user action = rebuild now. Force past the definition-hash cache
-    // so a refresh picks up new upstream data even when the query text is
-    // unchanged (parity with the dashboard data source Refresh button).
-    await materializeBinding(workspaceId, appId, bindingId, { force: true });
-    setMaterializing(false);
-  }, [workspaceId, appId, bindingId, materializeBinding]);
-
-  const bindingCache = binding?.cache;
-  const handleScheduleChange = useCallback(
-    (schedule: MaterializationScheduleValue) => {
-      if (!workspaceId) return;
-      updateBinding(appId, bindingId, { materializationSchedule: schedule });
-      void persistApp(workspaceId, appId);
-    },
-    [workspaceId, appId, bindingId, updateBinding, persistApp],
-  );
-
-  const handlePreviewSnapshot = useCallback(async () => {
-    if (!bindingCache?.parquetUrl) return;
-    setPreviewingSnapshot(true);
-    const startedAt = Date.now();
-    try {
-      const result = await previewParquetArtifact(bindingCache.parquetUrl);
-      setPreview({
-        results: result.rows,
-        executedAt: new Date().toISOString(),
-        resultCount: result.totalRows,
-        executionTime: Date.now() - startedAt,
-        fields: result.fields.map(f => f.name),
-      });
-      setViewMode("table");
-    } catch (e) {
-      setPreview({
-        results: [
-          {
-            error:
-              e instanceof Error
-                ? e.message
-                : "Failed to preview the materialized data",
-          },
-        ],
-        executedAt: new Date().toISOString(),
-        resultCount: 0,
-      });
-    } finally {
-      setPreviewingSnapshot(false);
-    }
-  }, [bindingCache?.parquetUrl]);
-
-  if (!appEntity) {
-    if (appLoadError) {
-      return (
-        <EntityLoadErrorState
-          error={appLoadError}
-          entityLabel="app"
-          onRetry={() => {
-            if (workspaceId) void fetchApp(workspaceId, appId);
-          }}
-        />
-      );
-    }
-    return <EntityLoadingState label="Loading data source…" />;
-  }
-  if (!binding) {
-    return (
-      <EntityLoadErrorState
-        error={missingEntityError("data source")}
-        entityLabel="data source"
-        detail="This data source no longer exists in this app."
-      />
-    );
-  }
-
-  const cache = binding.cache;
-
-  const historyItems: MaterializationHistoryItem[] = (cache?.history ?? []).map(
-    (run, i) => ({
-      id: `${run.at}-${i}`,
-      status: run.status === "ready" ? "ready" : "error",
-      at: run.at,
-      rowCount: run.rowCount ?? null,
-      durationMs: run.durationMs ?? null,
-      error: run.error ?? null,
-    }),
-  );
-
-  const isParquet = binding.materialization === "parquet";
-
-  // Show the dbt link picker when the workspace has dbt projects, or when the
-  // binding is already linked (so the link stays visible/unlinkable even if
-  // its project was deleted).
-  const showDbtLink = dbtProjects.length > 0 || Boolean(binding.dbtProjectId);
-  const linkedDbtProjectMissing = Boolean(
-    binding.dbtProjectId &&
-      dbtProjectsLoaded &&
-      !dbtProjects.some(p => p._id === binding.dbtProjectId),
-  );
-  const dbtLinkedWithoutToken = Boolean(
-    binding.dbtProjectId && !containsDbtSchemaToken(binding.code),
-  );
-
-  // Deliberately compact: the data-source toolbar shares one row with the
-  // Console's run/save/connection controls and hardly fits on normal screens.
-  // No caption labels; the Live/Materialized explanation lives in a tooltip on
-  // the toggle itself, and the dbt token warning is an icon, not a chip.
-  const leadingControls = (
-    <>
-      <Tooltip
-        title={
-          <Box sx={{ p: 0.5 }}>
-            <Typography variant="caption" display="block" sx={{ mb: 0.5 }}>
-              <b>Live</b> — the query runs against the connection on every read.
-              Best for small, always-fresh lookups.
-            </Typography>
-            <Typography variant="caption" display="block">
-              <b>Materialized</b> — the query is snapshotted to a Parquet file
-              (stored like dashboards) and loaded into DuckDB in the browser, so
-              the app can run fast analytical SQL. Click <b>Refresh</b> to
-              rebuild the snapshot from source.
-            </Typography>
-          </Box>
-        }
-      >
-        <ToggleButtonGroup
-          size="small"
-          exclusive
-          value={binding.materialization}
-          onChange={(_e, value) => {
-            if (value && workspaceId) {
-              updateBinding(appId, bindingId, { materialization: value });
-              void persistApp(workspaceId, appId);
-            }
-          }}
-        >
-          <ToggleButton value="live">Live</ToggleButton>
-          <ToggleButton value="parquet">Materialized</ToggleButton>
-        </ToggleButtonGroup>
-      </Tooltip>
-      {showDbtLink && (
-        <>
-          {/* NOTE: deliberately NOT wrapped in a Tooltip — a tooltip anchored
-              on the Select stays visible while its menu is open and covers the
-              first menu items. The info icon carries the explanation. */}
-          <Select
-            size="small"
-            variant="outlined"
-            displayEmpty
-            value={binding.dbtProjectId ?? ""}
-            onChange={e => handleDbtLinkChange(e.target.value)}
-            sx={{ fontSize: "0.72rem", height: 26, minWidth: 90, ml: 0.5 }}
-            renderValue={value => {
-              if (!value) return "dbt: not linked";
-              return `dbt: ${
-                dbtProjects.find(p => p._id === value)?.name ??
-                (linkedDbtProjectMissing ? "deleted project" : "…")
-              }`;
-            }}
-          >
-            <MenuItem value="">Not linked</MenuItem>
-            {dbtProjects.map(p => (
-              <MenuItem key={p._id} value={p._id}>
-                {p.name}
-              </MenuItem>
-            ))}
-          </Select>
-          <Tooltip
-            title={
-              <Box sx={{ p: 0.5 }}>
-                <Typography variant="caption" display="block" sx={{ mb: 0.5 }}>
-                  Link this data source to a dbt project to write
-                  environment-agnostic SQL: use the{" "}
-                  <code>{"{{ dbt_schema }}"}</code> token instead of hardcoding
-                  a schema (e.g.{" "}
-                  <code>{"SELECT * FROM {{ dbt_schema }}.fct_orders"}</code>).
-                </Typography>
-                <Typography variant="caption" display="block">
-                  The token resolves to the project&apos;s prod environment for
-                  published apps and materialized snapshots, and the app preview
-                  gains an environment switcher to point the draft (including
-                  materialized bindings) at a dev schema.
-                </Typography>
-              </Box>
-            }
-          >
-            <InfoIcon
-              size={15}
-              strokeWidth={1.5}
-              style={{ opacity: 0.6, cursor: "help" }}
-            />
-          </Tooltip>
-          {dbtLinkedWithoutToken && (
-            <Tooltip
-              title={
-                'Add the {{ dbt_schema }} token to the query (e.g. "FROM ' +
-                '{{ dbt_schema }}.my_model") to enable the schema switcher ' +
-                "in the app preview."
-              }
-            >
-              <Box
-                sx={{
-                  display: "flex",
-                  alignItems: "center",
-                  color: "warning.main",
-                }}
-              >
-                <InfoIcon size={15} strokeWidth={1.5} />
-              </Box>
-            </Tooltip>
-          )}
-        </>
-      )}
-    </>
-  );
-
-  const headerExtras = (
-    <DataSourceMaterializationControls
-      leadingControls={leadingControls}
-      showMaterializeControls={isParquet}
-      buildStatus={cache?.parquetBuildStatus ?? null}
-      rowCount={cache?.rowCount ?? null}
-      builtAtMs={
-        cache?.parquetBuiltAt ? Date.parse(cache.parquetBuiltAt) : null
-      }
-      dataFreshnessTtlMs={
-        binding.materializationSchedule?.dataFreshnessTtlMs ?? null
-      }
-      onMaterialize={() => void handleMaterialize()}
-      materializing={materializing}
-      canPreview={cache?.parquetBuildStatus === "ready" && !!cache?.parquetUrl}
-      onPreviewSnapshot={() => void handlePreviewSnapshot()}
-      previewing={previewingSnapshot}
-      schedule={
-        binding.materializationSchedule ?? defaultMaterializationSchedule(false)
-      }
-      onScheduleChange={handleScheduleChange}
-      scheduleCaption="This schedule refreshes only this app data source."
-      history={historyItems}
-    />
-  );
-
-  return (
-    <Box sx={{ height: "100%", display: "flex", flexDirection: "column" }}>
-      <EntityBreadcrumbs tabId={tabId} />
-
-      <Box sx={{ flexGrow: 1, minHeight: 0 }}>
-        <PanelGroup direction="vertical" style={{ height: "100%" }}>
-          <Panel defaultSize={55} minSize={20}>
-            <Console
-              variant="data-source"
-              consoleId={tabId}
-              initialContent={binding.code}
-              filePath={bindingFilePath(binding.name, binding.language)}
-              onExecute={handleExecute}
-              onSave={handleSave}
-              isExecuting={running}
-              connectionId={binding.connectionId}
-              onDatabaseChange={connectionId => {
-                if (workspaceId) {
-                  updateBinding(appId, bindingId, { connectionId });
-                  void persistApp(workspaceId, appId);
-                }
-              }}
-              onDatabaseNameChange={(databaseId, databaseName) => {
-                if (workspaceId) {
-                  updateBinding(appId, bindingId, { databaseId, databaseName });
-                  void persistApp(workspaceId, appId);
-                }
-              }}
-              databaseId={binding.databaseId}
-              databaseName={binding.databaseName}
-              headerExtras={headerExtras}
-            />
-          </Panel>
-          <PanelResizeHandle
-            style={{
-              height: 4,
-              background: "var(--mui-palette-divider, #ddd)",
-            }}
-          />
-          <Panel defaultSize={45} minSize={10}>
-            <ResultsTable
-              results={preview}
-              viewMode={viewMode}
-              onViewModeChange={setViewMode}
-            />
-          </Panel>
-        </PanelGroup>
-      </Box>
-    </Box>
-  );
-}
