+++ app/src/components/ShareDialog.tsx
@@ -71,14 +71,12 @@ export interface ShareDialogProps {
 const RESOURCE_LABEL: Record<ShareResourceType, string> = {
   dashboard: "dashboard",
   console: "console",
-  app: "app",
 };
 
 /** Public links are only available where snapshot data exists. */
 const SUPPORTS_PUBLIC: Record<ShareResourceType, boolean> = {
   dashboard: true,
   console: false,
-  app: true,
 };
 
 // Stable fallback so the Zustand selector never returns a fresh `[]` per call
@@ -379,23 +377,6 @@ export default function ShareDialog({
         return result;
       });
     }
-  };
-
-  const handleLiveToggle = async (allowLiveQueries: boolean) => {
-    if (!workspaceId || !resourceId) return;
-    await run(async () => {
-      const result = await updatePublicShare(
-        resourceType,
-        workspaceId,
-        resourceId,
-        { allowLiveQueries },
-      );
-      if (result.ok && result.publicShare) {
-        setPublicShare(result.publicShare);
-        onSharingChanged?.({ publicShare: result.publicShare });
-      }
-      return result;
-    });
   };
 
   const handleSetPassword = async () => {
@@ -1045,41 +1026,6 @@ export default function ShareDialog({
                       }}
                     />
                   ))}
-
-                {resourceType === "app" && (
-                  <Box
-                    sx={{
-                      display: "flex",
-                      alignItems: "flex-start",
-                      gap: 1,
-                      mt: 0.5,
-                      pt: 1,
-                      borderTop: "1px solid",
-                      borderColor: "divider",
-                    }}
-                  >
-                    <Box sx={{ flex: 1, minWidth: 0 }}>
-                      <Typography variant="caption" sx={{ fontWeight: 500 }}>
-                        Live data
-                      </Typography>
-                      <Typography
-                        variant="caption"
-                        color="text.secondary"
-                        sx={{ display: "block" }}
-                      >
-                        {publicShare.allowLiveQueries
-                          ? "Viewers run this app's published queries live (read-only)"
-                          : "Viewers see the last snapshot only"}
-                      </Typography>
-                    </Box>
-                    <Switch
-                      size="small"
-                      checked={!!publicShare.allowLiveQueries}
-                      disabled={!canManage || busy}
-                      onChange={e => void handleLiveToggle(e.target.checked)}
-                    />
-                  </Box>
-                )}
               </Box>
             )}
           </>
