+++ app/src/lib/entity-icons.ts
@@ -93,9 +93,6 @@ export const TAB_KIND_ICONS = {
   dashboard: ChartPie,
   "dashboard-data-source": Database,
   "table-data": Table,
-  app: AppWindow,
-  "app-file": FileCode,
-  "app-binding": Database,
   "app-v2": AppWindowMac,
   "app-v2-file": FileCode,
   "app-v2-diff": GitCompare,
