+++ app/src/lib/entity-labels.ts
@@ -19,9 +19,6 @@ export const TAB_KIND_ENTITY_LABELS = {
   dashboard: "dashboard",
   "dashboard-data-source": "data source",
   "table-data": "table",
-  app: "app",
-  "app-file": "file",
-  "app-binding": "data source",
   "app-v2": "app",
   "app-v2-file": "file",
   "app-v2-diff": "diff",
