+++ app/src/lib/tab-routing.ts
@@ -15,7 +15,6 @@
  * test in `tab-routing.test.ts` keeps the two directions in sync.
  */
 import type { ConsoleTab, TabKind } from "../store/lib/types";
-import { appLocationToHostSearch } from "../app-runtime/app-location";
 
 /** Encode a path that may contain slashes, keeping the slashes readable. */
 export function encodePathSegments(path: string): string {
@@ -41,9 +40,6 @@ export const TAB_DEEP_LINK_PATTERNS = {
   dashboard: /^\/d\/([a-zA-Z0-9-]+)\/?$/,
   "dashboard-data-source": /^\/d\/([a-zA-Z0-9-]+)\/data\/([a-zA-Z0-9_-]+)/,
   "table-data": /^\/t\/([a-zA-Z0-9-]+)\/([^/]+)\/([^/]+)\/?$/,
-  app: /^\/a\/([a-zA-Z0-9-]+)\/?$/,
-  "app-file": /^\/a\/([a-zA-Z0-9-]+)\/file\/(.+)$/,
-  "app-binding": /^\/a\/([a-zA-Z0-9-]+)\/data\/([a-zA-Z0-9_-]+)/,
   // Apps v2 (git-backed, experimental) — /a2 so it can't collide with /a.
   "app-v2": /^\/a2\/([a-zA-Z0-9-]+)\/?$/,
   "app-v2-file": /^\/a2\/([a-zA-Z0-9-]+)\/file\/(.+)$/,
@@ -99,28 +95,6 @@ export function tabUrlPath(tabId: string, tab: ConsoleTab): string | null {
         `/t/${tab.connectionId}/${encodeURIComponent(schema || "public")}` +
         `/${encodeURIComponent(table)}${query ? `?${query}` : ""}`
       );
-    }
-    case "app": {
-      const appId = tab.metadata?.appId as string | undefined;
-      if (!appId) return null;
-      // The running app projects its own location (path + query) onto the
-      // host URL: query params stay readable, the app pathname rides in the
-      // reserved `_path` param. The bare `/a/:appId` pathname is unchanged, so
-      // the deep-link pattern still matches.
-      const appLocation = tab.metadata?.appLocation as string | undefined;
-      return `/a/${appId}${appLocationToHostSearch(appLocation)}`;
-    }
-    case "app-file": {
-      const appId = tab.metadata?.appId as string | undefined;
-      const path = tab.metadata?.path as string | undefined;
-      return appId && path
-        ? `/a/${appId}/file/${encodePathSegments(path)}`
-        : null;
-    }
-    case "app-binding": {
-      const appId = tab.metadata?.appId as string | undefined;
-      const bindingId = tab.metadata?.bindingId as string | undefined;
-      return appId && bindingId ? `/a/${appId}/data/${bindingId}` : null;
     }
     case "app-v2": {
       // Prefer the slug: it is the app's folder name in the workspace repo,
