+++ app/src/store/lib/types.ts
@@ -17,9 +17,6 @@ export type TabKind =
   | "dashboard"
   | "dashboard-data-source"
   | "table-data"
-  | "app"
-  | "app-file"
-  | "app-binding"
   | "app-v2"
   | "app-v2-file"
   | "app-v2-diff"
