+++ packages/agent-tools/src/capabilities/app-capabilities.ts
@@ -29,187 +29,9 @@ const define = (
 
 export const APP_CAPABILITIES = [
   // ── Orientation / reads ─────────────────────────────────────────────────
-  define({
-    name: "list_open_apps",
-    pack: "app-orient",
-    risk: "read",
-    surfaces: ALL_AGENT_SURFACES,
-    resultKind: "data",
-  }),
-  define({
-    name: "get_app_state",
-    pack: "app-orient",
-    risk: "read",
-    surfaces: ALL_AGENT_SURFACES,
-    resultKind: "data",
-  }),
-  define({
-    name: "app_search",
-    pack: "app-orient",
-    risk: "read",
-    surfaces: ALL_AGENT_SURFACES,
-    resultKind: "data",
-  }),
-  define({
-    name: "app_read_resource",
-    pack: "app-orient",
-    risk: "read",
-    surfaces: ALL_AGENT_SURFACES,
-    resultKind: "data",
-  }),
-  define({
-    name: "app_read_file",
-    pack: "app-orient",
-    risk: "read",
-    surfaces: ALL_AGENT_SURFACES,
-    resultKind: "data",
-  }),
   // ── Files / dependencies ────────────────────────────────────────────────
-  define({
-    name: "create_app",
-    pack: "app-edit",
-    risk: "write",
-    requiredGrant: "artifact-write",
-    surfaces: ALL_AGENT_SURFACES,
-    resultKind: "artifact",
-  }),
-  define({
-    name: "app_write_file",
-    pack: "app-edit",
-    risk: "write",
-    requiredGrant: "artifact-write",
-    surfaces: ALL_AGENT_SURFACES,
-    resultKind: "artifact",
-  }),
-  define({
-    name: "app_edit_file",
-    pack: "app-edit",
-    risk: "write",
-    requiredGrant: "artifact-write",
-    surfaces: ALL_AGENT_SURFACES,
-    resultKind: "artifact",
-  }),
-  define({
-    name: "app_rename_file",
-    pack: "app-edit",
-    risk: "write",
-    requiredGrant: "artifact-write",
-    surfaces: ALL_AGENT_SURFACES,
-    resultKind: "artifact",
-  }),
-  define({
-    name: "app_delete_file",
-    pack: "app-edit",
-    risk: "destructive",
-    requiredGrant: "artifact-write",
-    surfaces: ALL_AGENT_SURFACES,
-    resultKind: "artifact",
-  }),
-  define({
-    name: "app_add_dependency",
-    pack: "app-edit",
-    risk: "write",
-    requiredGrant: "artifact-write",
-    surfaces: ALL_AGENT_SURFACES,
-    resultKind: "artifact",
-  }),
-  define({
-    name: "app_remove_dependency",
-    pack: "app-edit",
-    risk: "destructive",
-    requiredGrant: "artifact-write",
-    surfaces: ALL_AGENT_SURFACES,
-    resultKind: "artifact",
-  }),
   // ── Data bindings / materialization ─────────────────────────────────────
-  define({
-    name: "app_create_data_binding",
-    pack: "app-data",
-    risk: "write",
-    requiredGrant: "artifact-write",
-    surfaces: ALL_AGENT_SURFACES,
-    resultKind: "artifact",
-  }),
-  define({
-    name: "app_update_data_binding",
-    pack: "app-data",
-    risk: "write",
-    requiredGrant: "artifact-write",
-    // Scheduling folded in from app_set_binding_schedule: the schedule leg
-    // keeps that tool's schedule-write gate, applied only when the input
-    // actually touches the schedule.
-    inputConditionalGrants: [
-      {
-        grant: "schedule-write",
-        behavior: "changing a binding's materialization schedule",
-        appliesTo: input =>
-          typeof input === "object" &&
-          input !== null &&
-          (input as { materializationSchedule?: unknown })
-            .materializationSchedule !== undefined,
-      },
-    ],
-    surfaces: ALL_AGENT_SURFACES,
-    resultKind: "artifact",
-  }),
-  define({
-    name: "app_delete_data_binding",
-    pack: "app-data",
-    risk: "destructive",
-    requiredGrant: "artifact-write",
-    surfaces: ALL_AGENT_SURFACES,
-    resultKind: "artifact",
-  }),
-  define({
-    // Deprecated alias of app_update_data_binding({ materialization }); kept
-    // registered for external MCP clients. Deferred out of the in-product
-    // working set (see DEFERRED_BUILTIN_TOOL_DOMAINS).
-    name: "app_set_binding_materialization",
-    pack: "app-data",
-    risk: "write",
-    requiredGrant: "artifact-write",
-    surfaces: ALL_AGENT_SURFACES,
-    resultKind: "artifact",
-  }),
-  define({
-    // Executes the binding's saved query against the connection, so it needs
-    // the same query envelope as run_console / sql_execute_query.
-    name: "materialize_binding",
-    pack: "app-data",
-    risk: "write",
-    requiredGrant: "artifact-write",
-    surfaces: ALL_AGENT_SURFACES,
-    resultKind: "run",
-    requiresQueryAccess: true,
-  }),
-  define({
-    // Deprecated alias of app_update_data_binding({ materializationSchedule });
-    // kept registered for external MCP clients. Deferred out of the
-    // in-product working set (see DEFERRED_BUILTIN_TOOL_DOMAINS).
-    name: "app_set_binding_schedule",
-    pack: "app-data",
-    risk: "write",
-    requiredGrant: "schedule-write",
-    surfaces: ALL_AGENT_SURFACES,
-    resultKind: "artifact",
-  }),
   // ── Versions ────────────────────────────────────────────────────────────
-  define({
-    name: "app_save_version",
-    pack: "app-versions",
-    risk: "write",
-    requiredGrant: "artifact-write",
-    surfaces: ALL_AGENT_SURFACES,
-    resultKind: "artifact",
-  }),
-  define({
-    name: "app_restore_version",
-    pack: "app-versions",
-    risk: "write",
-    requiredGrant: "artifact-write",
-    surfaces: ALL_AGENT_SURFACES,
-    resultKind: "artifact",
-  }),
   // ── Client-only UI effects ──────────────────────────────────────────────
   define({
     name: "open_app",
